requirements.txt
Django>=5.0,<5.2
djangorestframework>=3.15
django-q2>=1.7
google-api-python-client>=2.130
google-auth>=2.30
google-auth-httplib2>=0.2
cloudinary>=1.40
requests>=2.32
